Darwin: Sync type of num_iso_packets fields
* num_iso_packets was a "size_t" in "struct darwin_transfer_priv" and an
  "int" in "struct libusb_transfer".  The field is now an int in the two
  structures
* Fixes warning
  os/darwin_usb.c: In function ‘submit_iso_transfer’:
  os/darwin_usb.c:1334: warning: comparison between signed and unsigned

+++ libusb/os/darwin_usb.h
@@ -147,7 +147,7 @@ struct darwin_device_handle_priv {
 struct darwin_transfer_priv {
   /* Isoc */
   IOUSBIsocFrame *isoc_framelist;
-  size_t num_iso_packets;
+  int num_iso_packets;
 
   /* Control */
 #if !defined (LIBUSB_NO_TIMEOUT_DEVICE)
